Add test tile set
I need to add a test tile set for creating the ice cave map.  The art
team has not provided us with any real assets to work with, so this will
have to do for the implementation aspects of map generation.

It's really just a button mash on the keyboard; I suspect there should
be comments in the Tiled files denoting which characters are which
entities for that map.