Fix map render
It turns out I was using an Isometric map renderer rather than an
orthogonal one.  What does this mean?  Isometric maps are ones meant to
be viewed at an angle rather than top-down.  Orthogonal maps are viewed
top-down---orthogonal to the plane.

In this case, it really is all how you look at it.