Organize BSPlayer class structure for graphics
                            VOLATILE COMMIT

In the old code, BSPlayer was an ordered mess of closely related objects
that handled each state's animation, its sprite sheet, and the
individual frames of that sprite-sheet (as TextureRegion arrays --- good
idea @dogmelons).  Now, each of those qualities are encapsulated in a
BSAnimation internal class.

I also copied and added some notes to this class's `draw` method.  See
the comment therein for more information, but long story short I think
it can be trivialized to one line now.

None of this code has been tested and it will likely cause the project
to not build.  Keep this in mind.