Organize BSPlayer class structure for graphics
                            VOLATILE COMMIT

I missed a few things on that last commit, and I'm too lazy to simply
amend.  (I also, admittedly, fear messing it up---this is probably not
called for, but I'm still pretty new at this.  In time.)

For each BSAnimation there was a method that I missed (of the pattern
`Set[...]Animations`) that seems to act as something of a constructor
for the animation.  (Can @dogmelons confirm?)  I included this and also
a `time` field that I missed previously.

This commit accounts for most of the functionality provided by the
`Player` class on the `master` branch.  All that remains to do is the
switching of directions, and I'm not positive who will be taking care of
that yet.  My plan is for it to be handled by `move`, which will
maintain some state in BSGameState (`isMoving`) to handle appropriate
flow through the animations.  A state machine will have to be coded in,
but that's simple enough to do.  (Yay, Theory of Computation!)