Directional pad visible
It so happens that the camera was taking the sprite batch and
converting *all* sprites in the batch to the world's coordinate system;
since the controls aren't part of the world, it makes more sense to have
a 'controls' batch.

The control wasn't visible because 'white' is a transparency color and
the d-pad had several hundred transparent pixels to show off. :)